fix osx build

+++ source/slang/slang-reflection-json.cpp
@@ -550,7 +550,7 @@ static void emitReflectionSize(PrettyWriter& writer, size_t size)
 {
     size == SLANG_UNBOUNDED_SIZE ? writer << "\"unbounded\""
     : size == SLANG_UNKNOWN_SIZE ? writer << "\"unknown\""
-                                 : writer << size;
+                                 : writer << (uint64_t)size;
 }
